Fix a bug in the support for browsers requiring polyfills.
The web UI requires web-components in order work, and not all
browsers support them fully yet. This was supposed to be worked
around by using the 'webcomponents-lite.min.js' library, but
that work-around was not properly implemented.

Specifically, there was no sequencing of the calls to the
'Polymer' function to ensure that they happened after the
web components library was loaded. The result was that the
'Polymer' function would be called before the web components
library was loaded, that name would not yet be defined, and
as a result the UI would completely fail to load.

This change fixes that by not allowing the initial calls to
the 'Polymer' function to happen until after the web components
library has finished loading.