Simplified the diff code by storing blocks instead of lines.
This change reworks the object model of the diff parser to
make the code much simpler.

Specifically, the new model adds the concept of a diff-block,
which is the aggregation of all of the sequential diff lines
that have the same status (omitted, unchanged, changed).

This results in not needing to use different models for inline
diffs vs. side-by-side diffs, as we no longer need to add
'missing' lines to pad the two sides of the side-by-side view.